requirements.txt
pika==1.3.2
paho-mqtt~=2.1.0